feat: merge dashboard API Response Status widget for v2 and v4 APIs
